# pkg/apispec/openapi.yaml
openapi: 3.0.3
info:
  title: Sandbox0 API
  version: 0.1.0
  description: Public HTTP APIs exposed by internal-gateway.
servers:
  - url: https://api.sandbox0.ai
    description: Production
...
paths:
...
  /api/v1/sandboxes/{id}/contexts/{ctx_id}/restart:
    post:
      tags: [contexts]
      summary: Restart context
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts/{ctx_id}/restart
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - $ref: "#/components/parameters/ContextID"
      responses:
        "200":
          description: Context restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessContextResponse"
  /api/v1/sandboxes/{id}/contexts/{ctx_id}/input:
    post:
      tags: [contexts]
      summary: Send input to context
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts/{ctx_id}/input
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - $ref: "#/components/parameters/ContextID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextInputRequest"
      responses:
        "200":
          description: Input written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessWrittenResponse"
  /api/v1/sandboxes/{id}/contexts/{ctx_id}/exec:
    post:
      tags: [contexts]
      summary: Execute context input (sync)
      description: Sends input and blocks until the context completes or times out.
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts/{ctx_id}/exec
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - $ref: "#/components/parameters/ContextID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextInputRequest"
      responses:
        "200":
          description: Execution completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessContextExecResponse"
  /api/v1/sandboxes/{id}/contexts/{ctx_id}/resize:
    post:
      tags: [contexts]
      summary: Resize context PTY
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts/{ctx_id}/resize
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - $ref: "#/components/parameters/ContextID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResizeContextRequest"
      responses:
        "200":
          description: Resized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResizedResponse"
  /api/v1/sandboxes/{id}/contexts/{ctx_id}/signal:
    post:
      tags: [contexts]
      summary: Send signal to context
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts/{ctx_id}/signal
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - $ref: "#/components/parameters/ContextID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalContextRequest"
      responses:
        "200":
          description: Signal sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSignaledResponse"
  /api/v1/sandboxes/{id}/contexts/{ctx_id}/stats:
    get:
      tags: [contexts]
      summary: Get context stats
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts/{ctx_id}/stats
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - $ref: "#/components/parameters/ContextID"
      responses:
        "200":
          description: Context stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessContextStatsResponse"
  /api/v1/sandboxes/{id}/contexts/{ctx_id}/ws:
    get:
      tags: [contexts]
      summary: Context WebSocket (I/O)
      description: |
        Upgrades to WebSocket for streaming I/O.
        Client messages (JSON):
        - { "type": "input", "data": "ls\n", "request_id": "req-1" }
        - { "type": "resize", "rows": 24, "cols": 80 }
        - { "type": "signal", "signal": "INT" }
        Server messages (JSON):
        - { "type": "output", "source": "stdout", "data": "hello\n" }
      security:
        - bearerAuth: []
      x-upstream-service: procd
      x-upstream-path: /api/v1/contexts/{ctx_id}/ws
      x-websocket:
        inbound:
          $ref: "#/components/schemas/ContextWebSocketRequest"
        outbound:
          $ref: "#/components/schemas/ContextWebSocketResponse"
      parameters:
        - $ref: "#/components/parameters/SandboxID"
        - $ref: "#/components/parameters/ContextID"
      responses:
        "101":
          description: Switching Protocols
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SandboxID:
      name: id
      in: path
      required: true
      schema:
        type: string
    ContextID:
      name: ctx_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data: {}
      required: [success]
...
    SuccessWrittenResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                written:
                  type: boolean
...
    SuccessResizedResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                resized:
                  type: boolean
    SuccessSignaledResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                signaled:
                  type: boolean
...
    ResourceUsage:
      type: object
      properties:
        cpu_percent:
          type: number
          format: double
        memory_rss:
          type: integer
          format: int64
        memory_vms:
          type: integer
          format: int64
        open_files:
          type: integer
        thread_count:
          type: integer
        container_memory_usage:
          type: integer
          format: int64
        container_memory_limit:
          type: integer
          format: int64
        container_memory_working_set:
          type: integer
          format: int64
        io_read_bytes:
          type: integer
          format: int64
        io_write_bytes:
          type: integer
          format: int64
        memory_bytes:
          type: integer
          format: int64
...
    ProcessType:
      type: string
      enum: [repl, cmd]
...
    ContextResponse:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/ProcessType"
        language:
          type: string
        cwd:
          type: string
        env_vars:
          type: object
          additionalProperties:
            type: string
        running:
          type: boolean
        paused:
          type: boolean
        created_at:
          type: string
        output:
          type: string
        public_url:
          type: string
        exposed_port:
          type: integer
          format: int32
      required: [id, type, running, paused, created_at]
    ContextStatsResponse:
      type: object
      properties:
        context_id:
          type: string
        type:
          type: string
        language:
          type: string
        running:
          type: boolean
        paused:
          type: boolean
        usage:
          $ref: "#/components/schemas/ResourceUsage"
    ContextInputRequest:
      type: object
      properties:
        data:
          type: string
      required: [data]
    ContextExecResponse:
      type: object
      properties:
        output:
          type: string
      required: [output]
    ResizeContextRequest:
      type: object
      properties:
        rows:
          type: integer
          format: int32
        cols:
          type: integer
          format: int32
      required: [rows, cols]
    SignalContextRequest:
      type: object
      properties:
        signal:
          type: string
      required: [signal]
    SuccessContextResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ContextResponse"
...
    SuccessContextStatsResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ContextStatsResponse"
    SuccessContextExecResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ContextExecResponse"
...
    ContextWebSocketRequest:
      oneOf:
        - $ref: "#/components/schemas/ContextWebSocketInput"
        - $ref: "#/components/schemas/ContextWebSocketResize"
        - $ref: "#/components/schemas/ContextWebSocketSignal"
    ContextWebSocketInput:
      type: object
      properties:
        type:
          type: string
          enum: [input]
        data:
          type: string
        request_id:
          type: string
      required: [type]
    ContextWebSocketResize:
      type: object
      properties:
        type:
          type: string
          enum: [resize]
        rows:
          type: integer
          format: int32
        cols:
          type: integer
          format: int32
      required: [type, rows, cols]
    ContextWebSocketSignal:
      type: object
      properties:
        type:
          type: string
          enum: [signal]
        signal:
          type: string
      required: [type, signal]
    ContextWebSocketResponse:
      oneOf:
        - $ref: "#/components/schemas/ContextWebSocketOutput"
        - $ref: "#/components/schemas/ContextWebSocketDone"
    ContextWebSocketOutput:
      type: object
      properties:
        type:
          type: string
          enum: [output]
        source:
          type: string
          enum: [stdout, stderr, prompt]
        data:
          type: string
      required: [type, source, data]
    ContextWebSocketDone:
      type: object
      properties:
        type:
          type: string
          enum: [done]
        request_id:
          type: string
      required: [type, request_id]
...